+++ packages/@ember/-internals/metal/index.ts
@@ -70,8 +70,6 @@ export {
 } from './lib/observer';
 export { default as inject, DEBUG_INJECTION_FUNCTIONS } from './lib/injected_property';
 export { tagForProperty, tagForObject, markObjectAsDirty } from './lib/tags';
-export { tracked, TrackedDescriptor } from './lib/tracked';
-export { cached } from './lib/cached';
 export { createCache, getValue, isConst } from './lib/cache';
 
 export {
